% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_Lambda_EM.r
\name{lcphR}
\alias{lcphR}
\title{lcphR: Latent Class Proportional Hazards Regression}
\usage{
lcphR(
  dat,
  num_class,
  covx,
  tolEM = 1e-07,
  tolNewton = 1e-08,
  maxiterEM = 1000,
  maxiterNewton = 50,
  initial = "null",
  init.tau = NULL,
  varest = FALSE,
  numvar = "1st",
  traceplot = FALSE,
  verbose = TRUE,
  warmup = 200
)
}
\arguments{
\item{dat}{Input data.}

\item{num_class}{Number of latent classes.}

\item{covx}{Column names of baseline covariates.}

\item{tolEM}{Stopping criteria for the EM algorithm.}

\item{tolNewton}{Stopping criteria for the Newton-Raphson algorithm in the M-step.}

\item{maxiterEM}{Number of maximum iterations of the EM algorithm.}

\item{maxiterNewton}{Number of maximum iterations of the Newton-Raphson algorithm.}

\item{initial}{Initialization setting, either 'kmeans', 'null', 'random'. Ignored if probability vectors are provided in 'init.tau'.}

\item{init.tau}{Provided membership probability based on prior knowledge, a n times c matrix where rows represent subjects and columns represent classes.}

\item{varest}{TRUE or FALSE, indicating whether conducting variance estimation or not.}

\item{numvar}{If varest is TRUE, this indicates the numerical method for profile likelihood based variance estimation, either "1st", "1st_rich", or "2nd", indicating 1st order numerical differentiation, 1st order differentiation based on Richardson extrapolation, and 2nd order numerical differentiation, respectively.}

\item{traceplot}{If TRUE, plot observed data log likelihood and complete data log likelihood versus number of iterations.}

\item{verbose}{If TRUE, print notifications when point estimation or variance estimation starts.}

\item{warmup}{Number of warm-up iterations needed for non-informative initial value. Default is 200.}
}
\value{
A list with point estimates (alpha, zeta), variance estimates [ASE (observed log likelihood based), nASESr (profile log likelihood based)].
}
\description{
Conduct latent class analysis with time-to-event data.
}
\examples{

set.seed(21227)
dat <- simulation(n=1000)
lcfit <- lcphR(dat,num_class=2,covx=c('Xcov1','Xcov2'),tolEM=1e-3,maxiterEM=100,varest=TRUE,traceplot=TRUE,initial='kmeans')

}
\references{
Fei, T., Hanfelt, J., & Peng, L. (2022). Latent Class Analysis with Semi-parametric Proportional Hazards Submodel for Time-to-event Data. arXiv preprint arXiv:2202.00775.
}
\author{
Teng Fei. Email: feit1@mskcc.org
}
