% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.r
\name{simulation}
\alias{simulation}
\title{simulation: simulate a dataset for analysis}
\usage{
simulation(n = 1000)
}
\arguments{
\item{n}{sample size of the simulated dataset.}
}
\description{
Simulate a dataset according to the scenario (I) in the manuscript.
}
\examples{

dat <- simulation(n=1000)

}
\references{
Fei, Hanfelt and Peng. Latent Class Analysis with Semi-parametric Proportional Hazards Submodel for Time-to-event Data (submitted).
}
\author{
Teng Fei. Email: tfei@emory.edu
}
